Render entrances.
New layer "entrances", with entrance and access tags.

Render entrances starting at z=17. The "main" and "staircase" entrances
are rendered with a square marker, the "service" entrance with a circle
and other entrances with a disc.

Three levels of marker brightness reflect the access permissions of the
entrance. Darker color means more permissive access. The "main"
entrances are dark by default. The access tag modifies the brightness
such that access="yes" and "permissive" set the marker to darkest and
access=no to the lightest colour. Markers with other access tag values
have intermediate brightness.

Exits, such as entrance=exit and entrance=emergency are not rendered.

Address label text-halo-radius is increased at z>=18 to improve
readability of labels on top of entrance markers.